+++ Source/SelfService/Web/microservice/dataTable.tsx
@@ -12,7 +12,7 @@ import { DataGridPro, GridColDef, GridValueGetterParams, GridRenderCellParams } 
 import { Box, Paper, Tooltip } from '@mui/material';
 
 const capitalize = (str: string) => {
-    if(str.length < 1){
+    if (str.length < 1) {
         return str;
     }
 
@@ -21,7 +21,7 @@ const capitalize = (str: string) => {
 
 const sortByRuntimeVersion = (params: GridValueGetterParams) => {
     const runtimeVersion = params.row.edit?.extra?.runtimeImage?.replace(/dolittle\/runtime:/gi, '');
-    if(typeof runtimeVersion !== 'string'){
+    if (typeof runtimeVersion !== 'string') {
         return 'N/A';
     }
 
@@ -67,6 +67,7 @@ type DataTableProps = {
 export const DataTable = ({ application, environment, microservices }: DataTableProps) => {
     const history = useHistory();
     const [rows, setRows] = useState<(MicroserviceObject | undefined)[]>([]);
+    const [loadingRows, setLoadingRows] = useState<boolean>(true);
 
     const getMicroserviceStatus = useCallback(async (microserviceId: string) => {
         const status = await getPodStatus(application.id, environment, microserviceId);
@@ -81,8 +82,11 @@ export const DataTable = ({ application, environment, microservices }: DataTable
                 ...microservice,
                 phase: status[0]?.phase
             } as MicroserviceObject;
-        }))
-            .then(setRows);
+        })).then(data => {
+            setRows(data);
+            setLoadingRows(false);
+        });
+
     }, []);
 
     const customUrlFieldSort = (v1, v2, param1, param2) => {
@@ -150,12 +154,13 @@ export const DataTable = ({ application, environment, microservices }: DataTable
                 columns={columns}
                 disableColumnMenu
                 hideFooter
+                loading={loadingRows}
                 autoHeight={true}
                 headerHeight={46}
                 getRowHeight={() => 'auto'}
-                loading={!rows}
                 disableSelectionOnClick
                 onRowClick={(params) => onTableRowClick(params.row.id)}
-            /></Box>
+            />
+        </Box>
     );
 };