+++ Source/SelfService/Web/microservice/podLogScreen.tsx
@@ -3,31 +3,39 @@
 
 import React, { useState, useEffect } from 'react';
 
+import { LoadingSpinner } from '@dolittle/design-system/atoms/LoadingSpinner/LoadingSpinner';
+
 import { Typography } from '@mui/material';
 
 import { getPodLogs, HttpResponsePodLog } from '../api/api';
 
 export const PodLogScreen = ({ applicationId, podName, containerName }: any) => {
     const [data, setData] = useState({ logs: '' } as HttpResponsePodLog);
+    const [loading, setLoading] = useState(true);
 
     useEffect(() => {
         getPodLogs(applicationId, podName, containerName).then(data => {
             setData(data);
+            setLoading(false);
             return;
         });
     }, []);
 
-    if (!data.logs) {
-        return (
-            <Typography variant="body2" sx={{ pl: 7.5, py: 1 }}>
-                There are no logs printed for this microservice.
-            </Typography>
-        );
-    }
+    if (loading) {
+        return <LoadingSpinner />;
+    };
 
     return (
-        <pre style={{ padding: '0.5rem 0.625rem 1.1875rem 3rem', margin: '0', whiteSpace: 'pre-wrap', wordWrap: 'break-word' }}>
-            {data.logs}
-        </pre>
+        <>
+            {!data.logs ?
+                <Typography variant="body2" sx={{ pl: 7.5, py: 1 }}>
+                    There are no logs printed for this microservice.
+                </Typography> :
+
+                <pre style={{ padding: '0.5rem 0.625rem 1.1875rem 3rem', margin: '0', whiteSpace: 'pre-wrap', wordWrap: 'break-word' }}>
+                    {data.logs}
+                </pre>
+            }
+        </>
     );
 };