+++ Source/SelfService/Web/microservice/purchaseOrder/create.tsx
@@ -1,40 +1,28 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 import React from 'react';
 import { useHistory } from 'react-router-dom';
-import {
-    Grid,
-    Button,
-    Typography
-} from '@mui/material';
-import { TabPanel } from '../../utils/materialUi';
-import { Guid } from '@dolittle/rudiments';
-import { useSnackbar } from 'notistack';
 
+import { Grid, Button, Typography } from '@mui/material';
+
+import { Guid } from '@dolittle/rudiments';
+import { Tabs } from '@dolittle/design-system/atoms/Tabs/Tabs';
+import { AlertBox } from '@dolittle/design-system/atoms/AlertBox/AlertBox';
 
 import { savePurchaseOrderMicroservice } from '../../stores/microservice';
+import { HttpResponseApplication } from '../../api/application';
 import { MicroservicePurchaseOrder } from '../../api/index';
 import { getLatestRuntimeInfo } from '../../api/api';
-import { useGlobalContext } from '../../stores/notifications';
 import { Configuration } from './configuration';
-import { Tab, Tabs } from '../../theme/tabs';
-import { HttpResponseApplication } from '../../api/application';
 
-type Props = {
+type CreateProps = {
     application: HttpResponseApplication;
     environment: string;
 };
 
-export const Create: React.FunctionComponent<Props> = (props) => {
+export const Create = ({ application: { id, customerId }, environment }: CreateProps) => {
     const history = useHistory();
-    const { enqueueSnackbar } = useSnackbar();
-
-    const _props = props!;
-    const application = _props.application;
-    const environment = _props.environment;
-
-
-    const { setNotification } = useGlobalContext();
 
     const microserviceId = Guid.create().toString();
     const headImage = 'dolittle/integrations-m3-purchaseorders:latest';
@@ -42,13 +30,13 @@ export const Create: React.FunctionComponent<Props> = (props) => {
 
     const ms: MicroservicePurchaseOrder = {
         dolittle: {
-            applicationId: application.id,
-            customerId: application.customerId,
+            applicationId: id,
+            customerId,
             microserviceId,
         },
         name: '',
         kind: 'purchase-order-api',
-        environment: _props.environment,
+        environment,
         extra: {
             ingress: {
                 path: '/api/webhooks',
@@ -61,72 +49,43 @@ export const Create: React.FunctionComponent<Props> = (props) => {
         },
     };
 
-
-    const handleChange = (event: React.ChangeEvent<{}>, newValue: number) => {
-        // TODO replace with enqueue
-        setNotification(
-            'Health Status only available after microservice created',
-            'error'
-        );
+    const handleCancel = () => {
+        const href = `/microservices/application/${id}/${environment}/create`;
+        history.replace(href);
     };
 
     const _onSave = (ms: MicroservicePurchaseOrder): void => {
         // TODO handle exception (maybe move to wait)
         savePurchaseOrderMicroservice(ms).then((data) => {
-            const href = `/microservices/application/${application.id}/${environment}/view/${microserviceId}?firstTime=1`;
+            const href = `/microservices/application/${id}/${environment}/view/${microserviceId}?firstTime=1`;
             history.push(href);
         }).catch(reason => console.log(reason));
     };
 
     return (
-        <Grid
-            container
-            direction='column'
-            justifyContent='flex-start'
-        >
-            <Typography variant='h1' my={2}>Create purchase order API</Typography>
-
-            <Grid
-                container
-                spacing={3}
-
-            >
-                <Grid item xs={10}>
-                    <Tabs
-                        value={0}
-                        onChange={handleChange}
-                    >
-                        <Tab label='Configuration' />
-                        <Tab label='Health Status' />
-                    </Tabs>
-                </Grid>
-
-                <Grid
-                    item xs={2}
-                >
-                    <Grid
-                        container
-                        direction='row'
-                        justifyContent='space-around'
-                        alignItems='flex-end'
-                    >
-                        <Button
-                            onClick={() => {
-                                const href = `/microservices/application/${application.id}/${environment}/create`;
-                                history.replace(href);
-                            }}
-                        >
-                            cancel
-                        </Button>
-                    </Grid>
-                </Grid>
+        <>
+            <Grid container direction='column'>
+                <Typography variant='h1' my={2}>Create purchase order API</Typography>
+                <Button sx={{ alignSelf: 'flex-end' }} onClick={handleCancel}>Cancel</Button>
             </Grid>
 
-
-
-            <TabPanel value={0} index={0}>
-                <Configuration onSave={_onSave} microservice={ms} />
-            </TabPanel>
-        </Grid >
+            <Tabs
+                sx={{ mb: 3.25 }}
+                tabs={[
+                    {
+                        label: 'Configuration',
+                        render: () => <Configuration onSave={_onSave} microservice={ms} />,
+                    },
+                    {
+                        label: 'Health Status',
+                        render: () => <AlertBox
+                            severity='error'
+                            title='Health Status is only available after microservice is created.'
+                            sx={{ maxWidth: 520 }}
+                        />,
+                    }
+                ]}
+            />
+        </>
     );
 };