+++ Source/SelfService/Web/screens/applicationsScreen/applicationsScreen.tsx
@@ -27,7 +27,7 @@ export const ApplicationsScreen = () => {
     const history = useHistory();
     const { enqueueSnackbar } = useSnackbar();
 
-    const [data, setData] = useState([] as ShortInfoWithEnvironment[]);
+    const [applicationInfos, setApplicationInfos] = useState([] as ShortInfoWithEnvironment[]);
     const [loaded, setLoaded] = useState(false);
     const [canCreateApplication, setCanCreateApplication] = useState(false);
     const { setCurrentEnvironment } = useGlobalContext();
@@ -39,7 +39,7 @@ export const ApplicationsScreen = () => {
         ]).then(values => {
             const response = values[0] as HttpResponseApplications;
             setCanCreateApplication(response.canCreateApplication);
-            setData(response.applications);
+            setApplicationInfos(response.applications);
             setLoaded(true);
         }).catch((error) => {
             console.log(error);
@@ -69,12 +69,12 @@ export const ApplicationsScreen = () => {
     return (
         <LoginWrapper>
             <Typography variant='h2' sx={styles}>
-                Select Your Application & Environment
+                { applicationInfos.length > 0 ? 'Select Your Application & Environment' : 'There are no Applications' }
             </Typography>
 
             <CreateButton onCreate={handleCreate} />
 
-            <ApplicationsList data={data} onChoose={onEnvironmentChoose} />
+            <ApplicationsList data={applicationInfos} onChoose={onEnvironmentChoose} />
 
             <ActionButtons />